+++ src/SKCamera.ts
@@ -1,0 +1,3 @@
+export default class SKCamera extends THREE.Camera{
+
+}
