--- src/WonderCamera.ts
@@ -1,3 +1,0 @@
-export default class WonderCamera extends THREE.Camera{
-
-}
