+++ src/objects/MultiMaterialObject.ts
