+++ src/objects/Sofa.ts
