+++ tsconfig.json
@@ -40,21 +40,24 @@
         "./typings/threejs/three-vreffect.d.ts",
         "./typings/threejs/three.d.ts",
         "./typings/tsd.d.ts",
-        "./src/CustomShaderRotatingCube.ts",
+        "./typings/tween.js/tween.js.d.ts",
+        "./src/event/click.ts",
+        "./src/event/drag.ts",
         "./src/main.ts",
-        "./src/MultiMaterialObject.ts",
-        "./src/RotatingCube.ts",
-        "./src/Sofa.ts",
-        "./src/SofaShader.ts",
-        "./src/SofaShaderSphere.ts",
-        "./src/type.ts",
-        "./src/VaseShaderSphere.ts",
-        "./src/WonderCamera.ts",
-        "./src/WonderObjLoader.ts",
-        "./src/WonderOperator.ts",
-        "./src/WonderRenderer.ts",
-        "./src/WonderScene.ts",
-        "./src/WonderVR.ts"
+        "./src/objects/CustomShaderRotatingCube.ts",
+        "./src/objects/MultiMaterialObject.ts",
+        "./src/objects/RotatingCube.ts",
+        "./src/objects/Sofa.ts",
+        "./src/objects/SofaShader.ts",
+        "./src/objects/SofaShaderSphere.ts",
+        "./src/objects/VaseShaderSphere.ts",
+        "./src/SKCamera.ts",
+        "./src/SKObjLoader.ts",
+        "./src/SKongVR.ts",
+        "./src/SKOperator.ts",
+        "./src/SKRenderer.ts",
+        "./src/SKScene.ts",
+        "./src/type.ts"
     ],
     "exclude": [],
     "atom": {
